+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ### bug fixes
 
+- fixed computation of Jacobian center for `ngmix.KObservation`s with mixed
+  even-odd dimensions
 - the total shape amplitude `g` is now reset properly when using the
   `Shape.set_g1g2` method
 - fixed large rounding errors in some shape conversions in `ngmix.shape`
@@ -14,6 +16,7 @@
   * `ngmix.shape`
   * `ngmix.jacobian`
   * `ngmix.moments`
+  * `ngmix.observation`
 
 ### deprecated/removed